+++ CHANGELOG.md
@@ -23,7 +23,7 @@ If you like this project and find it useful, please consider giving it a star on
   <img src="bmc-button.svg" alt="Buy me a coffee" width="120">
 </a>
 
-## [1.3.13] - 2025-09-??
+## [1.4.0] - 2025-10-15
 
 ### Added
 
@@ -32,10 +32,14 @@ If you like this project and find it useful, please consider giving it a star on
 
 ### Changed
 
+- [package]: Require matterbridge 3.3.0.
+- [index]: Updated to new signature PlatformMatterbridge.
+- [platform]: Updated to new signature PlatformMatterbridge.
+- [platform]: Added whiteList and blackList.
+- [platform]: Typed TestPlatformConfig.
 - [package]: Updated dependencies.
+- [package]: Bumped package to automator version 2.0.7
 - [workflows]: Ignore any .md anywhere in all workflows.
-- [package]: Updated to Automator v. 2.0.7.
-- [package]: Required matterbridge 3.2.9.
 
 <a href="https://www.buymeacoffee.com/luligugithub">
   <img src="bmc-button.svg" alt="Buy me a coffee" width="80">